image_generator/requirements.txt
Django==3.1.7
djangorestframework==3.12.2
gunicorn==20.0.4
matplotlib==3.3.4
numpy==1.20.1
Pillow==8.1.2
python-decouple==3.4
